lambda/enterprise/pyproject.toml
[project]
name = "mizpos-enterprise-android-manager"
version = "0.1.0"
description = "Android Enterprise MDM管理用Lambda Function"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.0",
    "fastapi>=0.115.0",
    "mangum>=0.19.0",
    "pydantic>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
]
